requirements.txt
Flask
pymongo
PyJWT
python-dotenv
cryptography
argon2-cffi
numpy
Werkzeug
